refactor: create di.factory.support package
