+++ nextstep-di/src/main/java/nextstep/di/factory/BeanFactory.java
@@ -1,5 +1,8 @@
 package nextstep.di.factory;
 
+import nextstep.di.factory.support.BeanFactoryUtils;
+import nextstep.di.factory.support.Beans;
+import nextstep.di.factory.support.ReflectionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
